+++ raaghu-components/rds-comp-reviews/rds-comp-reviews.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./rds-comp-reviews.scss";
-import { renderReviewStyle } from "./review-styles";
+import { renderReviewStyle } from "./rds-comp-review-styles";
 import Grid from "@mui/material/Grid";
 import Box from "@mui/material/Box";
